Move board-page components
